Minor debugging of function inputs.
